Allow using '~' in projects file to indicate home directory.
This is a partial solution to making projects portable: rather than
hardcoding the entire path in .smt/projects, only the path relative
to the home directory is hardcoded.
This commit simply adds `os.path.expanduser` in the relevant places,
allowing paths can be specified with '~' to indicate the home directory.
It does not provide a mechanism to change how they are written to
the project file; to make the project file portable, it needs to be edited
by hand. If the project file is re-saved (e.g. with smt configure),
the changes are overwritten.