Drop Python 2; change to `src` layout; vendorize parameters
The `src` layout in particular makes it easier to vendorize `sumatra`
into another project such as `smttask`.
Vendorizing allows us to patch packages as needed without making
installation complicated for the user. The one caveat is that
vendored packages cannot be installed alongside the originals, since
they are placed in the same location in `site-packages`.