Workaround: Disable checking that main file is in the project’s repo
If a project depends on multiple repos, it might be the case that `main_file`
is tracked with a different repo than `working_copy`. The best solution would probably
be to allow specifying multiple repos, and check that `main_file` is in one of them.
However currently Sumatra only allows specifying one repo.
As a workaround, we currently disable `check_file_under_version_control`
and trust that the user specified the appropriate repo when configuring the project.